+++ lib/chromium-phases/chromium-phases.js
@@ -3,15 +3,13 @@
 
 var register = require('../../core/phase-register.js');
 var types = require('../../core/types');
-var Executable = require('./executable');
-var options = require('../../core/options');
 
+var options = require('./chromium-options');
+var Executable = require('./executable');
 
 register({name: 'buildChromium', input: types.string, output: types.string, arity: '1:1', async: true},
   function(data, tags, cb) {
-    var chromiumBuildConfig = options.chromiumBuildConfig || 'Release';
-    var chromiumBuildTarget = options.chromiumBuildTarget || 'chrome';
-    var ninja = new Executable('ninja', [ '-C', 'out/' + chromiumBuildConfig, chromiumBuildTarget ]);
+    var ninja = new Executable('ninja', [ '-C', 'out/' + options.buildConfig, options.buildTarget ]);
     return ninja.run().then(function(code) {
       if (code) {
         var errorMessage = '"' + ninja + '" returned a non-zero exit code "' + code + '".\nThis is likely a Chromium build error.';
