+++ lib/chromium-phases/executable.js
@@ -1,10 +1,6 @@
 var Promise = require('bluebird');
-var options = require('../../core/options');
+var options = require('./chromium-options');
 var spawn = require('child_process').spawn;
-
-var chromiumPath = options.chromium;
-if (!chromiumPath)
-  throw new Error('Path to Chromium repo (--chromium=[path/to/cr/src]) is required.');
 
 function Executable(cmd, args) {
   this.cmd = cmd;
@@ -17,7 +13,7 @@ Executable.prototype.toString = function() {
 
 Executable.prototype.run = function() {
   var child = spawn(this.cmd, this.args, {
-    cwd: chromiumPath,
+    cwd: options.path,
     stdio: 'ignore',
   });
   child.on('error', function(err) {
